Add support for displaying the ISA string on RISC-V CPUs

+++ src/common/printing.c
@@ -135,7 +135,7 @@ void ffPrintCharTimes(char c, uint32_t times)
 
     char str[32];
     memset(str, c, sizeof(str)); //2 instructions when compiling with AVX2 enabled
-    for(uint32_t i = sizeof(str); i <= times; i += sizeof(str))
+    for(uint32_t i = sizeof(str); i <= times; i += (uint32_t)sizeof(str))
         fwrite(str, 1, sizeof(str), stdout);
     uint32_t remaining = times % sizeof(str);
     if(remaining > 0)
